Add first conditional is able to display after moving conditionals to a new file
