+++ src/Components/App/App.tsx
@@ -6,22 +6,13 @@ import { Route, Switch } from 'react-router-dom'
 import mockData from '../../MockData/mock-data.json'
 // import { data } from 'cypress/types/jquery'
 import callData from '../Api/ApiCalls'
-import fetchRecordings from '../Api/ApiCalls'
-
+import { FetchConditional } from '../../FetchConditonal/FetchConditional'
 
 type AppState = {
   count: number
   currentCnt: string
   currentType: string
   searchResults: Array<{}>
-}
-
-type ApiResponse = {
-  recordings: Recording[]
-}
-
-type Recording = {
-  type: string
 }
 
 class App extends Component<{}, AppState> {
@@ -34,26 +25,12 @@ class App extends Component<{}, AppState> {
 
 fetchResults = (event: React.MouseEvent<HTMLButtonElement>, formState: {selectedCnt : string, selectedType : string}) => {
   event.preventDefault()
-  // const filteredData = mockData.recordings.filter(data => 
-  //      data.cnt === formState.selectedCnt && data.type.includes(formState.selectedType)) 
-  //   console.log(filteredData)
-  // this.setState({searchResults: filteredData})  
-    console.log(formState)
-  const combinedParameters = `${formState.selectedCnt}-${formState.selectedType}`
-  switch (combinedParameters) {
-    case 'United States-flight-song':
-      fetchRecordings('https://xeno-canto.org/api/2/recordings?query=cnt:%22=United%20States')
-      .then((data: ApiResponse) => {
-        const filtered = data.recordings.filter(recording => recording.type === 'flight call')
-        console.log(filtered)
-        this.setState({ searchResults: filtered})
+  
+      FetchConditional(formState.selectedCnt, formState.selectedType)
+      .then(data => {
+        console.log(data)
+        this.setState({ searchResults: data})
       })
-      break;
-      default:
-        console.log('still works')
-      }
-      
-
   }
 
 render() {
